refactor: move execution-state to internal type
