+++ packages/core/src/new-api/internal/reconciliation/futures/reconcileReadEventArgument.ts
@@ -1,5 +1,5 @@
-import { ReadEventArgumentExecutionState } from "../../../types/execution-state";
 import { ReadEventArgumentFuture } from "../../../types/module";
+import { ReadEventArgumentExecutionState } from "../../types/execution-state";
 import { ExecutionStateResolver } from "../execution-state-resolver";
 import { ReconciliationContext, ReconciliationFutureResult } from "../types";
 import { fail } from "../utils";
