+++ packages/core/src/new-api/internal/types/execution-engine.ts
@@ -1,5 +1,6 @@
-import { ExecutionStateMap } from "../../types/execution-state";
 import { IgnitionModule, IgnitionModuleResult } from "../../types/module";
+
+import { ExecutionStateMap } from "./execution-state";
 
 export interface ExecutionEngineState {
   batches: string[][];
