+++ packages/core/test/new-api/reconciliation/futures/reconcileArtifactContractAt.ts
@@ -7,7 +7,7 @@ import {
   ExecutionStateMap,
   ExecutionStatus,
   StaticCallExecutionState,
-} from "../../../../src/new-api/types/execution-state";
+} from "../../../../src/new-api/internal/types/execution-state";
 import { FutureType } from "../../../../src/new-api/types/module";
 import { assertSuccessReconciliation, reconcile } from "../helpers";
