+++ packages/core/test/new-api/reconciliation/futures/reconcileReadEventArgument.ts
@@ -5,7 +5,7 @@ import {
   DeploymentExecutionState,
   ExecutionStatus,
   ReadEventArgumentExecutionState,
-} from "../../../../src/new-api/types/execution-state";
+} from "../../../../src/new-api/internal/types/execution-state";
 import { FutureType } from "../../../../src/new-api/types/module";
 import { assertSuccessReconciliation, reconcile } from "../helpers";
