+++ packages/core/test/new-api/reconciliation/reconciler.ts
@@ -4,7 +4,7 @@ import { defineModule } from "../../../src/new-api/define-module";
 import {
   DeploymentExecutionState,
   ExecutionStatus,
-} from "../../../src/new-api/types/execution-state";
+} from "../../../src/new-api/internal/types/execution-state";
 import { FutureType } from "../../../src/new-api/types/module";
 
 import { assertSuccessReconciliation, reconcile } from "./helpers";
